refactor: organize navigation component files
