+++ frontend/src/App.tsx
@@ -5,7 +5,7 @@ import { AuthProvider } from '@/contexts/AuthContext'
 import { useAuth } from '@/hooks/useAuth'
 import { useCacheValidation } from '@/hooks/useCacheValidation'
 import { useTheme } from '@/hooks/useTheme'
-import Navigation from '@/components/Navigation'
+import Navigation from '@/components/navigation/Navigation'
 import DashboardPage from '@/dashboard/DashboardPage'
 import AccountsPage from '@/accounts/AccountsPage'
 import AccountDetailPage from '@/accounts/detail/AccountDetailPage'
