+++ frontend/src/components/navigation/components/AnimatedMobileMenuIcon.tsx
