+++ frontend/src/components/navigation/components/NavigationBrand.tsx
