+++ frontend/src/components/navigation/components/NavigationUserProfile.tsx
