+++ frontend/src/components/navigation/hooks/useMobileNavigationEffects.ts
