+++ frontend/src/components/navigation/types/navigationTypes.ts
