+++ frontend/src/components/navigation/utils/navigationLabels.ts
